requirements.txt
pygame==2.0.0
pyserial==3.4
